+++ src/main/java/omnet/tactics/DecreaseDimmer.java
@@ -11,7 +11,7 @@ public class DecreaseDimmer extends FailableTactic {
 
 	// set statically for now
 	private static double failChance = 0.05;
-	private static double latency = 1;
+	private static long latency = 1;
 	
 	private String server;
 	
@@ -67,7 +67,7 @@ public class DecreaseDimmer extends FailableTactic {
 	}
 
 	@Override
-	public double getTime() {
+	public long getTime() {
 		return latency;
 	}
